+++ day-5/aoc2023-day-5/src/main/java/org/aoc2023/model/entityMap/EntityMapConfig.java
@@ -1,4 +1,4 @@
 package org.aoc2023.model.entityMap;
 
-public record EntityMapConfig(int source, int destination, int range) {
+public record EntityMapConfig(long source, long destination, long range) {
 }
